apps/backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
# psycopg2-binary==2.9.9  # Commented out - using SQLite initially

# Trading and Financial
python-binance==1.0.19
pandas==2.1.3
numpy==1.25.2
ta==0.10.2  # Technical analysis library
ccxt==4.1.77  # Cryptocurrency exchange library

# AI and Machine Learning
openai==1.3.7  # For OpenAI API (optional)
langchain==0.0.350  # For LLM integration
langchain-community==0.0.10
pydantic==2.5.0

# Data Processing and Analysis
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Real-time and Async
websockets==12.0
aiohttp==3.9.1
httpx==0.25.2
redis==5.0.1
aioredis==2.0.1

# Scheduling and Background Tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Configuration and Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development and Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3